requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
django-nocaptcha-recaptcha==0.0.20
gunicorn==20.0.4
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
django-phonenumber-field==5.0.0
django_unique_slugify==1.01
phonenumbers==8.12.19
gspread~=3.7.0
oauth2client~=4.1.3